src/lambdas/etl/requirements.txt
pandas
fastapi
numpy
uuid==1.30
scipy
psycopg2
pyarrow
python-dotenv
scipy
psycopg2-binary
boto3
